% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{classify_new_ts_taxonomy}
\alias{classify_new_ts_taxonomy}
\title{Classifies a ts based on the tstaxonomyr R package.}
\usage{
classify_new_ts_taxonomy(ts, ts_taxonomy = "v1")
}
\arguments{
\item{ts}{Either a vector, time series or data.frame object representing
a time series is allowed.}

\item{ts_taxonomy}{Either 'v1' or 'v2'. v1 uses the default time series
taxonomy and v2 the feature selected ts taxonomy of the tstaxonomyr package.}
}
\value{
A data frame, containing all scaled feature factor values.
If the input \code{ts} is not a vector, ts or data.frame object
an error message is returned.
}
\description{
This is a function to classify a time series data by the tstaxonomyr R
package. It requires a time series, vector or a data.frame object which
represents a multivariate time series. The time series object is classified
on several (statistical) features of the ts taxonomy from the
'tstaxonomyr' R package.
}
\examples{
classify_new_ts_taxonomy(ts = datasets::BJsales)
}
