% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fc_data_preparation}
\alias{fc_data_preparation}
\title{Prepares a data frame object for classification ml method}
\usage{
fc_data_preparation(ts_fc_data, class_label)
}
\arguments{
\item{ts_fc_data}{A time series or data frame object.}

\item{class_label}{A character value containing a column name of the
inputed data frame \code{ts_fc_data}.}
}
\value{
A data frame object containg the prepared data.
If the input is not a time series or data frame object, an error message
is returned.
}
\description{
This is a function to prepares a data frame object for the classification
ml process. All date cols are deleted. Also, all character and factor cols
are transformed by binarization through one hot encoding.
As input is only required an object from the class data.frame
Otherwise the function returns an error message.
}
\examples{
library(mlbench)
fc_data_preparation(train_ann = mlbench::PimaIndiansDiabetes,
class_label = 'diabetes')
}
