% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitted_ts_taxonomy_c50_model}
\alias{fitted_ts_taxonomy_c50_model}
\title{A fitted classification c5.0 rule model from the c50 R package}
\format{fitted c50 model from the c50 R package The method is
called 'c50'.}
\usage{
fitted_ts_taxonomy_c50_model
}
\description{
A fitted classification xgboost model from the caret R packe based on the
two .R files: "ts_taxonomy_results" and "ts_fc_evaluation_results"
from this package. The "ts_taxonomy_results" are applied as training features
and the "ts_fc_evaluation_results" as prediction label. Thus the model is
able to predict for new "ts_taxonomy_results" data the best performing
forecasting method out of the seven possible from the
"ts_fc_evaluation_results" file.
}
\keyword{datasets}
