% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_ml_models.R
\name{train_ann}
\alias{train_ann}
\title{Generates a classification ann trainings model.}
\usage{
train_ann(n_round = 10, cv_nfold = 10, tune_length = 10,
  ts_taxonomy = "v1")
}
\arguments{
\item{n_round}{Number of cross-validation rounds.}

\item{cv_nfold}{Number of folds of cross-validation.}

\item{tune_length}{Number of model tuning intervals.}

\item{ts_taxonomy}{Either 'v1' or 'v2'. v1 uses the default time series
taxonomy and v2 the feature selected ts taxonomy of the tstaxonomyr package.}
}
\value{
The best fitted ann classification model.
}
\description{
This is a function to generate a ann classification model for the
1000 time series from the tstaxonomyr R package. The best fitted ann model is
identified based on the k-fold \code{cv_nfold} cross-validation and the
number of rounds \code{n_round}. The method 'nnet' from 'caret' package is
fitted and afterwards returned. As input is only required the \code{n_round},
\code{cv_nfold} and \code{tune_length} with a number between 1 and 100. Also,
for \code{ts_taxonomy}  only 'v1' for the basic taxonomy or 'v2' for the
ligther feature selected taxonomy of the tstaxonomyr R package are allowed.
Otherwise the function returns an error message.
}
\examples{
fitted_model <- train_ann()
}
