% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_fc_evaluation_results}
\alias{ts_fc_evaluation_results}
\title{A data frame of 1000 time series data forecasting method evaluation.}
\format{A data frame with 1000 elements.
Each row represents a time series object with it's best performing
forecasting method. The following columns are provided:
\describe{
  \item{ts_name}{The name of the time series from the tstaxonomyr package}
  \item{best_fc_method}{The best performing forecasting method}
}}
\usage{
ts_fc_evaluation_results
}
\description{
A data frame containing 1000 different univariate and
multivariate time series data from the tstaxonomyr R package. For each one
is given the best performing forecasting method out of the seven
possible ones: 'random walk', 'exponentiel smoothing', 'arima', 'svm',
'ann', 'xgboost' and 'cart'.
}
\keyword{datasets}
