% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_fs_taxonomy_results}
\alias{ts_fs_taxonomy_results}
\title{A data frame of 1000 different classified time series data.}
\format{A data frame with 1000 elements.
Each row represents a time series object that is classified on the following
15 (statistical features) from the tstaxonomyr R package representing the
columns of the data frame. This time the time series data was classified on
the feature selection reduced taxonomy:
\describe{
  \item{Trend}{The trend factor of the time series}
  \item{Autocorrelation}{The autocorrelation factor of the time series}
  \item{Seasonality}{The seasonality factor of the time series}
  \item{Periodicity}{The periodicity factor of the time series}
  \item{Chaos}{The chaos factor of the time series}
  \item{Entropy}{The entropy factor of the time series}
  \item{TurningPoints}{The number of turning points of the time series}
  \item{PartialAutocorrelation}{The partial autocorrelation factor of
  the time series}
  \item{Outliers}{The number of outliers of the time series}
  \item{StepChanges}{The step changes factor of the time series}
  \item{Peaks}{The number of peaks of the time series}
  \item{DurbinWatsonTest}{The dw test factor of the time series}
  \item{calculate_determination_coefficient}{The R^2 factor of the time
  series}
  \item{NumberOfAttributes}{The number of features of the time series}
  \item{ts_name}{The name of the time series from the tstaxonomyr package}
}}
\source{
tstaxonomyr R package \url{https://github.com/mowomoyela/tstaxonomyr}
}
\usage{
ts_fs_taxonomy_results
}
\description{
A data frame containing 1000 different classified univariate and
multivariate time series data from the tstaxonomyr R package.
Each row represents one single time series that is classified on XX different
(statistical) features which are listed in the following item description.
This time the time series data was classified on
the feature selection reduced taxonomy. All features/columns are factor
values of the different classification classes.
}
\keyword{datasets}
