% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_taxonomy_results}
\alias{ts_taxonomy_results}
\title{A data frame of 1000 different classified time series data.}
\format{A data frame with 1000 elements.
Each row represents a time series object that is classified on the following
(statistical features) from the tstaxonomyr R package representing the
columns of the data frame:
\describe{
  \item{Skewness}{The skewness factor of the time series}
  \item{Kurtosis}{The kurtosis factor of the time series}
  \item{Trend}{The trend factor of the time series}
  \item{Autocorrelation}{The autocorrelation factor of the time series}
  \item{Mean}{The mean factor of the time series}
  \item{StandardDeviation}{The sd factor of the time series}
  \item{NumberOfObservations}{The length of the time series}
  \item{NonLinearity}{The non-linearity factor of the time series}
  \item{Seasonality}{The seasonality factor of the time series}
  \item{Periodicity}{The periodicity factor of the time series}
  \item{Chaos}{The chaos factor of the time series}
  \item{Entropy}{The entropy factor of the time series}
  \item{SelfSimilarity}{The self-similarity factor of the time series}
  \item{feature_dtwdistance}{The dynamic time warping distance to other
  time series blocks}
  \item{TurningPoints}{The number of turning points of the time series}
  \item{PartialAutocorrelation}{The partial autocorrelation factor of
  the time series}
  \item{Variance}{The variance factor of the time series}
  \item{Outliers}{The number of outliers of the time series}
  \item{StepChanges}{The step changes factor of the time series}
  \item{Peaks}{The number of peaks of the time series}
  \item{DurbinWatsonTest}{The dw test factor of the time series}
  \item{calculate_determination_coefficient}{The R^2 factor of the time
  series}
  \item{NumberOfAttributes}{The number of features of the time series}
  \item{ts_name}{The name of the time series from the tstaxonomyr package}
}}
\source{
tstaxonomyr R package \url{https://github.com/mowomoyela/tstaxonomyr}
}
\usage{
ts_taxonomy_results
}
\description{
A data frame containing 1000 different classified univariate and
multivariate time series data from the tstaxonomyr R package.
Each row represents one single time series that is classified on 24 different
(statistical) features which are listed in the following item description.
All features/columns are factor values of the different classification
classes.
}
\keyword{datasets}
